requirements.txt
# Core Dependencies - Python 3.13 Compatible
SpeechRecognition>=3.14.0
pyttsx3>=2.90
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0
pyautogui>=0.9.50

# Optional dependencies - may need alternatives
# spotipy>=1.22.0
# pygame>=2.5.0
# newsapi-python>=0.2.6
# pywin32>=306
